edit and delete button all language

+++ view/templates/product_overview.php
@@ -13,9 +13,7 @@
 
     if($type == $selected){
 
-    //echo "<a id=\"product-page\" href=\"".$this->build_url('index.php', array('action' => 'product_page'))."&id=$id \"></a>";
-
-   echo "<a id=\"product-page\" href='index.php?action=product_page&id=$id'>";
+    echo "<a id=\"product-page\" href=\"".$this->build_url('index.php', array('action' => 'product_page', 'id' => $id))." \">";
 
     echo "<div class=product-container>";
 
@@ -28,9 +26,9 @@
     </div>
     <?php echo "<p class='price'>CHF $price</p>";
 
-    /*if ($this->controller->isAdmin()) {
+    if ($this->controller->isAdmin()) {
     echo "<br><a href=\"index.php?action=edit_product&id=$id\">Edit</a> | <a href=\"index.php?action=delete_product&id=$id\">Delete</a><br/>";
-    }*/
+    }
     ?>
 </div>
 <?php
